refactor: renamed mongoldb to mongol
