+++ src/db.rs
@@ -1,3 +1,3 @@
-mod mongoldb;
+mod mongol;
 
-pub use mongoldb::*;
+pub use mongol::*;
