+++ src/db/mongol/model/message.rs
@@ -4,7 +4,7 @@ use bson::Bson;
 use mongodb::bson::{DateTime, Uuid};
 use serde::{Deserialize, Serialize};
 
-use crate::{db::mongoldb::{as_string, helper}, model::{message::{Message, MessageFlag}, error}};
+use crate::{db::mongol::{as_string, helper}, model::{message::{Message, MessageFlag}, error}};
 
 
 #[derive(Debug, Serialize, Deserialize)]
