+++ src/db/mongol/model/refresh_token.rs
@@ -1,7 +1,7 @@
 use bson::{Bson, DateTime, Uuid};
 use serde::{Deserialize, Serialize};
 
-use crate::{db::mongoldb::{as_string, helper, MongolHelper}, model::{error, refresh_token::{RefreshToken, RefreshTokenFlag}}};
+use crate::{db::mongol::{as_string, helper, MongolHelper}, model::{error, refresh_token::{RefreshToken, RefreshTokenFlag}}};
 
 
 #[derive(Debug, Serialize, Deserialize)]
