+++ src/db/mongol/model/user.rs
@@ -1,7 +1,7 @@
 use bson::Bson;
 use mongodb::bson::Uuid;
 use serde::{Serialize, Deserialize};
-use crate::{db::mongoldb::{as_string, helper}, model::{error, user::{User, UserFlag}}};
+use crate::{db::mongol::{as_string, helper}, model::{error, user::{User, UserFlag}}};
 
 #[derive(Debug, Serialize, Deserialize)]
 #[allow(clippy::pub_underscore_fields)]
