+++ src/db/mongol/repositories/user.rs
@@ -4,7 +4,7 @@ use futures_util::StreamExt;
 use mongodb::bson::{doc, from_document, Uuid};
 
 use crate::model::{error, Pagination, user::{User, UserFlag, UserRepository}};
-use crate::db::mongoldb::{helper, MongolDB, MongolUser, MongolUserVec};
+use crate::db::mongol::{helper, MongolDB, MongolUser, MongolUserVec};
 use crate::map_mongo_key_to_string;
 
 #[async_trait]
